refactor: rename dto to dtos in auth module
